requirements.txt
pyspark
google-generativeai
python-dotenv
streamlit
pandas
